AulaTest/CLI_CLIENTE/ClienteTO.java
package CLI_CLIENTE;

public class ClienteTO {
	private int id;
		private String nome;
		private String telefone;
		private String endereco;
		private String nasc;
	
		public String getTelefone() {
			return telefone;
		}
	
		public void setTelefone(String telefone) {
			this.telefone = telefone;
		}
	
		public String getEndereco() {
			return endereco;
		}
	
		public void setEndereco(String endereco) {
			this.endereco = endereco;
		}
	
		public String getNasc() {
			return nasc;
		}
	
		public void setNasc(String nasc) {
			this.nasc = nasc;
		}
	
		public int getId() {
			return id;
		}
	
		public void setId(int id) {
			this.id = id;
		}
	
		public String getNome() {
			return nome;
		}
	
		public void setNome(String nome) {
			this.nome = nome;
		}
		
		public String getNascimento(){
			return nasc;
		}
		
		public void setNascimento (String nascimento){
			this.nasc = nascimento;
		}
}
